// src/app/shared/model/BikePark.ts
/**
 * Api Documentation
 * Api Documentation
 *
 * OpenAPI spec version: 1.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import * as models from './models';

export interface BikePark {
  contact?: models.Contact;

  denumire?: string;

  descriere?: string;

  id?: number;

  nrLocuri?: number;

  telescaun?: boolean;

  user?: models.User;

}