// src/app/shared/model/Locatie.ts
/**
 * Api Documentation
 * Api Documentation
 *
 * OpenAPI spec version: 1.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import * as models from './models';

export interface Locatie {
  codPostal?: string;

  coordonate?: string;

  id?: number;

  latitude?: number;

  localitate?: string;

  longitude?: number;

  number?: string;

  provincie?: string;

  strada?: string;

  tara?: string;

}