// src/app/shared/model/Photo.ts
/**
 * Api Documentation
 * Api Documentation
 *
 * OpenAPI spec version: 1.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import * as models from './models';

export interface Photo {
  deleted?: boolean;

  id?: number;

  path?: string;

  publicId?: string;

  url?: string;

}